devices.json
The purpose of devices. JSON is to define and store configuration data, such as session keys, parameters, and other metadata, for a particular device. For tasks like provisioning devices, establishing secure communication, or integrating them into an IoT system, it provides a human-readable and machine-parsable structure for organizing and exchanging this data using the JSON format.Itiscontained in the expo file.